# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pâtisserie ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated

  # DASHBOARD
...
  /branches:
    get:
      operationId: getBranches
      tags: [branches]
      summary: List branches
      responses:
        "200":
          description: Branches list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: lowStock
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        username:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended, archived, invited]
        language:
          type: string
          enum: [fr, ar]
        roleId:
          type: ["integer", "null"]
        roleName:
          type: ["string", "null"]
        branchIds:
          type: array
          items:
            type: integer
        defaultBranchId:
          type: ["integer", "null"]
        posAccess:
          type: boolean
        adminAccess:
          type: boolean
        lastLogin:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - username
        - status
        - language
        - posAccess
        - adminAccess
        - branchIds
        - createdAt
...
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [shop, lab, warehouse, central]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        isMain:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - code
        - type
        - isActive
        - isMain
        - createdAt
    CreateBranchBody:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [shop, lab, warehouse, central]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        isMain:
          type: boolean
      required:
        - name
        - code
        - type
        - isActive
        - isMain
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        type:
          type: string
          enum: [finished, ingredient, packaging, semi_finished]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        unitId:
          type: integer
        unitName:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        alertQuantity:
          type: ["number", "null"]
        shelfLifeDays:
          type: ["integer", "null"]
        isManaged:
          type: boolean
        isSellable:
          type: boolean
        isPurchasable:
          type: boolean
        isFabricated:
          type: boolean
        totalStock:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - unitId
        - unitName
        - costPrice
        - sellingPrice
        - isManaged
        - isSellable
        - isPurchasable
        - isFabricated
        - totalStock
        - createdAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        type:
          type: string
          enum: [finished, ingredient, packaging, semi_finished]
        categoryId:
          type: ["integer", "null"]
        unitId:
          type: integer
        description:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        alertQuantity:
          type: ["number", "null"]
        shelfLifeDays:
          type: ["integer", "null"]
        isManaged:
          type: boolean
        isSellable:
          type: boolean
        isPurchasable:
          type: boolean
        isFabricated:
          type: boolean
      required:
        - name
        - type
        - unitId
        - costPrice
        - sellingPrice
        - isManaged
        - isSellable
        - isPurchasable
        - isFabricated
...